chore(tests): made tests synchronous and added winner endpoint test
